+++ CHANGELOG.md
@@ -9,7 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.5.32] 2022-03-08
 
+### Added
 
+- Serialization/Deserialization of **GridGroupLayout** component was added.
 
 ## [0.5.31] 2022-03-08