+++ CHANGELOG.md
@@ -7,11 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.5.32] 2022-03-08
+## [0.5.32] 2022-03-09
 
 ### Added
 
 - Serialization/Deserialization of **GridGroupLayout** component was added.
+
+### Fixed
+
+- The issue when no camera isn't marked as active was fixed. Now if there is no an active game camera the editor's one is used for that role.
 
 ## [0.5.31] 2022-03-08